// packages/types/src/generated/prisma/models/project_members.ts
export type project_membersSelect<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = runtime.Types.Extensions.GetSelect<
  {
    project_id?: boolean;
    user_id?: boolean;
    status?: boolean;
    created_by?: boolean;
    created_at?: boolean;
    updated_by?: boolean;
    updated_at?: boolean;
    project?: boolean | Prisma.projectsDefaultArgs<ExtArgs>;
    user?: boolean | Prisma.usersDefaultArgs<ExtArgs>;
    created_by_user?:
      boolean | Prisma.project_members$created_by_userArgs<ExtArgs>;
    updated_by_user?:
      boolean | Prisma.project_members$updated_by_userArgs<ExtArgs>;
  },
  ExtArgs['result']['project_members']
>;
// ...
export type project_membersOmit<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = runtime.Types.Extensions.GetOmit<
  | 'project_id'
  | 'user_id'
  | 'status'
  | 'created_by'
  | 'created_at'
  | 'updated_by'
  | 'updated_at',
  ExtArgs['result']['project_members']
>;
export type project_membersInclude<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = {
  project?: boolean | Prisma.projectsDefaultArgs<ExtArgs>;
  user?: boolean | Prisma.usersDefaultArgs<ExtArgs>;
  created_by_user?:
    boolean | Prisma.project_members$created_by_userArgs<ExtArgs>;
  updated_by_user?:
    boolean | Prisma.project_members$updated_by_userArgs<ExtArgs>;
};
// ...
export type $project_membersPayload<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = {
  name: 'project_members';
  objects: {
    project: Prisma.$projectsPayload<ExtArgs>;
    user: Prisma.$usersPayload<ExtArgs>;
    created_by_user: Prisma.$usersPayload<ExtArgs> | null;
    updated_by_user: Prisma.$usersPayload<ExtArgs> | null;
  };
  scalars: runtime.Types.Extensions.GetPayloadResult<
    {
      project_id: string;
      user_id: string;
      status: $Enums.RecordStatus;
      created_by: string | null;
      created_at: Date;
      updated_by: string | null;
      updated_at: Date;
    },
    ExtArgs['result']['project_members']
  >;
  composites: {};
};

export type project_membersGetPayload<
  S extends boolean | null | undefined | project_membersDefaultArgs,
> = runtime.Types.Result.GetResult<Prisma.$project_membersPayload, S>;
// ...
export type project_membersDefaultArgs<
  ExtArgs extends runtime.Types.Extensions.InternalArgs =
    runtime.Types.Extensions.DefaultArgs,
> = {
  /**
   * Select specific fields to fetch from the project_members
   */
  select?: Prisma.project_membersSelect<ExtArgs> | null;
  /**
   * Omit specific fields from the project_members
   */
  omit?: Prisma.project_membersOmit<ExtArgs> | null;
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.project_membersInclude<ExtArgs> | null;
};